Move transcoding to own module
